requirements.txt
Django>=3.0,<4.0
psycopg2-binary>=2.8
channels
channels_redis==2.4.2
django-background-tasks
